added routes to website
