+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/index.js":"1","/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/App.js":"2","/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/reportWebVitals.js":"3","/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/components/Snake.js":"4","/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/components/Food.js":"5"},{"size":500,"mtime":1611705394969,"results":"6","hashOfConfig":"7"},{"size":3165,"mtime":1611879546234,"results":"8","hashOfConfig":"7"},{"size":362,"mtime":1611705394970,"results":"9","hashOfConfig":"7"},{"size":437,"mtime":1611715169277,"results":"10","hashOfConfig":"7"},{"size":247,"mtime":1611715147388,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"1mlhvci",{"filePath":"15","messages":"16","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"14"},"/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/index.js",[],["23","24"],"/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/App.js",[],"/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/reportWebVitals.js",[],"/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/components/Snake.js",[],"/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/components/Food.js",[],{"ruleId":"25","replacedBy":"26"},{"ruleId":"27","replacedBy":"28"},"no-native-reassign",["29"],"no-negated-in-lhs",["30"],"no-global-assign","no-unsafe-negation"]
+[{"/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/index.js":"1","/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/App.js":"2","/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/reportWebVitals.js":"3","/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/components/Snake.js":"4","/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/components/Food.js":"5","/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/Nav.js":"6","/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/pages/About.js":"7","/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/pages/SnakeGame.js":"8"},{"size":500,"mtime":1611887248872,"results":"9","hashOfConfig":"10"},{"size":669,"mtime":1611887188357,"results":"11","hashOfConfig":"10"},{"size":362,"mtime":1611705394970,"results":"12","hashOfConfig":"10"},{"size":437,"mtime":1611715169277,"results":"13","hashOfConfig":"10"},{"size":247,"mtime":1611715147388,"results":"14","hashOfConfig":"10"},{"size":628,"mtime":1611888011422,"results":"15","hashOfConfig":"10"},{"size":153,"mtime":1611886324538,"results":"16","hashOfConfig":"10"},{"size":3194,"mtime":1611886807491,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"1mlhvci",{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"22"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/index.js",[],"/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/App.js",[],["35","36"],"/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/reportWebVitals.js",[],"/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/components/Snake.js",[],"/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/components/Food.js",[],"/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/Nav.js",[],"/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/pages/About.js",[],"/Users/matthew/Desktop/DigitalCrafts/Week 20/snake-game/src/pages/SnakeGame.js",[],{"ruleId":"37","replacedBy":"38"},{"ruleId":"39","replacedBy":"40"},"no-native-reassign",["41"],"no-negated-in-lhs",["42"],"no-global-assign","no-unsafe-negation"]
