+++ src/index.css
@@ -1,3 +1,27 @@
+/* NavBar */
+nav {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  background: rgb(0, 0, 0);
+  color: white;
+  padding-left: 40px;
+  border-radius: 10px;
+  opacity: 85%;
+  border-bottom: outset;
+}
+
+.nav-links {
+  width: 30%;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  list-style: none;
+  text-decoration: none;
+}
+
+
+/* Snake Game */
 .game-area {
   position: relative;
   margin: 50px auto;
