+++ src/pages/About.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function About() {
+    return (
+        <div>
+           <h1>About Me Page</h1>
+        </div>
+    )
+}
+
+export default About;
